requirements.txt
numpy~=1.26
pillow>=10.0,<11
opencv-python>=4.9,<5
scikit-image>=0.22,<0.25
pandas>=2.2,<3
matplotlib>=3.8,<4
tqdm>=4.66,<5
